Pain Train: +30% health from healers, -50% from hp packs
The pain train has no changes in VSH: Rewrite, so this is a proposed change to make it unique. It's kind of a reverse back scratcher, allowing you to be more safe with a medic, but pretty doomed without one. This does not affect concheror healing.

(this also capitalizes 'boost' in the bfb's description to make it consistent with the rest because it's just below)